Interface Functions

Index provider interface

Mandatory methods

SymbolicIndexingInterface.symbolic_container — Function

symbolic_container(indp)

Using indp, return an object that implements the index provider interface. In case indp itself implements the interface, indp can be returned as-is. All index provider interface methods fall back to calling the same method on symbolic_container(indp), so this may be used for trivial implementations of the interface that forward all calls to another object.

SymbolicIndexingInterface.is_variable — Function

is_variable(indp, sym)

Check whether the given sym is a variable in indp.

SymbolicIndexingInterface.variable_index — Function

variable_index(indp, sym, [i])

Return the index of the given variable sym in indp, or nothing otherwise. If constant_structure is false, this accepts the current time index as an additional parameter i.

SymbolicIndexingInterface.variable_symbols — Function

variable_symbols(indp, [i])

Return a vector of the symbolic variables being solved for in the index provider indp. If constant_structure(sys) == false this accepts an additional parameter indicating the current time index. The returned vector should not be mutated.

For types that implement Base.getindex with symbolic indices using this interface, the shorthand valp[solvedvariables] can be used as shorthand for valp[variable_symbols(sys)]. See: solvedvariables.

SymbolicIndexingInterface.is_parameter — Function

is_parameter(indp, sym)

Check whether the given sym is a parameter in indp.

SymbolicIndexingInterface.parameter_index — Function

parameter_index(indp, sym)

Return the index of the given parameter sym in indp, or nothing otherwise.

SymbolicIndexingInterface.parameter_symbols — Function

parameter_symbols(indp)

Return a vector of the symbolic parameters of the given index provider indp. The returned vector should not be mutated.

SymbolicIndexingInterface.is_independent_variable — Function

is_independent_variable(indp, sym)

Check whether the given sym is an independent variable in indp. The returned vector should not be mutated.

SymbolicIndexingInterface.independent_variable_symbols — Function

independent_variable_symbols(indp)

Return a vector of the symbolic independent variables of the given index provider indp.

SymbolicIndexingInterface.is_observed — Function

is_observed(indp, sym)

Check whether the given sym is an observed value in indp.

SymbolicIndexingInterface.default_values — Function

default_values(indp)

Return a dictionary mapping symbols in the index provider to their default value, if any. This includes parameter symbols. The dictionary must be mutable.

SymbolicIndexingInterface.is_time_dependent — Function

is_time_dependent(indp)

Check if indp has time as (one of) its independent variables.

SymbolicIndexingInterface.constant_structure — Function

constant_structure(indp)

Check if indp has a constant structure. Constant structure index providers do not change the number of variables or parameters over time.

SymbolicIndexingInterface.all_variable_symbols — Function

all_variable_symbols(indp)

Return a vector of variable symbols in the system, including observed quantities.

For types that implement Base.getindex with symbolic indices using this interface, The shorthand sys[allvariables] can be used as shorthand for valp[all_variable_symbols(indp)].

See: allvariables.

SymbolicIndexingInterface.all_symbols — Function

all_symbols(indp)

Return an array of all symbols in the index provider. This includes parameters and independent variables.

SymbolicIndexingInterface.solvedvariables — Constant

const solvedvariables = SolvedVariables()

This singleton is used as a shortcut to allow indexing of all solution variables (excluding observed quantities). It has a symbolic_type of ScalarSymbolic. See: variable_symbols.

SymbolicIndexingInterface.allvariables — Constant

const allvariables = AllVariables()

This singleton is used as a shortcut to allow indexing of all solution variables (including observed quantities). It has a symbolic_type of ScalarSymbolic. See all_variable_symbols.

Optional Methods

Observed equation handling

SymbolicIndexingInterface.observed — Function

observed(indp, sym, [states])

Return the observed function of the given sym in indp. The returned function should have the signature (u, p) -> [values...] where u and p is the current state and parameter vector, respectively. If istimedependent(indp) == true, the function should accept the current time t as its third parameter. If constant_structure(indp) == false, observed accepts a third parameter, which can either be a vector of symbols indicating the order of states or a time index, which identifies the order of states. This function does not need to be defined if is_observed always returns false. Thus, it is mandatory to always check is_observed before using this function.

See also: is_time_dependent, constant_structure

Value provider interface

Parameter indexing

SymbolicIndexingInterface.parameter_values — Function

parameter_values(valp)
parameter_values(valp, i)

Return an indexable collection containing the value of each parameter in valp. The two- argument version of this function returns the parameter value at index i. The two-argument version of this function will default to returning parameter_values(valp)[i].

If this function is called with an AbstractArray or Tuple, it will return the same array/tuple.

SymbolicIndexingInterface.set_parameter! — Function

set_parameter!(valp, val, idx)

Set the parameter at index idx to val for value provider valp. This defaults to modifying parameter_values(valp). If any additional bookkeeping needs to be performed or the default implementation does not work for a particular type, this method needs to be defined to enable the proper functioning of setp.

See: parameter_values

SymbolicIndexingInterface.finalize_parameters_hook! — Function

finalize_parameters_hook!(valp, sym)

This is a callback run one for each call to the function returned by setp which can be used to update internal data structures when parameters are modified. This is in contrast to set_parameter! which is run once for each parameter that is updated.

SymbolicIndexingInterface.getp — Function

getp(indp, sym)

Return a function that takes an value provider, and returns the value of the parameter sym. The value provider has to at least store the values of parameters in the corresponding index provider. Note that sym can be an index, symbolic variable, or an array/tuple of the aforementioned.

If sym is an array/tuple of parameters, then the returned function can also be used as an in-place getter function. The first argument is the buffer (must be an AbstractArray) to which the parameter values should be written, and the second argument is the value provider.

Requires that the value provider implement parameter_values. This function may not always need to be implemented, and has a default implementation for collections that implement getindex.

If the returned function is used on a timeseries object which saves parameter timeseries, it can be used to index said timeseries. The timeseries object must implement parameter_timeseries, parameter_values_at_time and parameter_values_at_state_time. The function returned from getp will can be passed Colon() (:) as the last argument to return the entire parameter timeseries for p, or any index into the parameter timeseries for a subset of values.

SymbolicIndexingInterface.setp — Function

setp(indp, sym)

Return a function that takes an index provider and a value, and sets the parameter sym to that value. Note that sym can be an index, a symbolic variable, or an array/tuple of the aforementioned.

Requires that the value provider implement parameter_values and the returned collection be a mutable reference to the parameter object. In case parameter_values cannot return such a mutable reference, or additional actions need to be performed when updating parameters, set_parameter! must be implemented.

SymbolicIndexingInterface.ParameterIndexingProxy — Type

struct ParameterIndexingProxy

This struct wraps any struct implementing the value provider and index provider interfaces. It allows getindex and setindex! operations to get/set parameter values. Requires that the wrapped type support getp and setp for getting and setting parameter values respectively.

Parameter timeseries

If a solution object saves a timeseries of parameter values that are updated during the simulation (such as by callbacks), it must implement the following methods to ensure correct functioning of getu and getp.

SymbolicIndexingInterface.parameter_timeseries — Function

parameter_timeseries(valp)

Return an iterable of time steps at which the parameter values are saved. This is only required for objects where is_timeseries(valp) === Timeseries() and the parameter values change during the simulation (such as through callbacks). By default, this returns [0].

See also: parameter_values_at_time.

SymbolicIndexingInterface.parameter_values_at_time — Function

parameter_values_at_time(valp, i)

Return an indexable collection containing the value of all parameters in valp at time index i. This is useful when parameter values change during the simulation (such as through callbacks) and their values are saved. i is the time index in the timeserie formed by these changing parameter values, obtained using parameter_timeseries.

By default, this function returns parameter_values(valp) regardless of i, and only needs to be specialized for timeseries objects where parameter values are not constant at all times. The resultant object should be indexable using parameter_values.

If this function is implemented, parameter_values_at_state_time must be implemented for getu to work correctly.

SymbolicIndexingInterface.parameter_values_at_state_time — Function

parameter_values_at_state_time(valp, i)

Return an indexable collection containing the value of all parameters in valp at time index i. This is useful when parameter values change during the simulation (such as through callbacks) and their values are saved. i is the time index in the timeseries formed by dependent variables (as opposed to the timeseries of the parameters, as in parameter_values_at_time).

By default, this function returns parameter_values(valp) regardless of i, and only needs to be specialized for timeseries objects where parameter values are not constant at all times. The resultant object should be indexable using parameter_values.

If this function is implemented, parameter_values_at_time must be implemented for getp to work correctly.

State indexing

SymbolicIndexingInterface.Timeseries — Type

struct Timeseries <: IsTimeseriesTrait end

Trait indicating a type contains timeseries data. This affects the behaviour of functions such as state_values and current_time.

See also: NotTimeseries, is_timeseries

SymbolicIndexingInterface.NotTimeseries — Type

struct NotTimeseries <: IsTimeseriesTrait end

Trait indicating a type does not contain timeseries data. This affects the behaviour of functions such as state_values and current_time. Note that if a type is NotTimeseries this only implies that it does not store timeseries data. It may still be time-dependent. For example, an ODEProblem only stores the initial state of a system, so it is NotTimeseries, but still time-dependent. This is the default trait variant for all types.

See also: Timeseries, is_timeseries

SymbolicIndexingInterface.is_timeseries — Function

is_timeseries(x) = is_timeseries(typeof(x))
is_timeseries(::Type)

Get the timeseries trait of a type. Defaults to NotTimeseries for all types.

See also: Timeseries, NotTimeseries

SymbolicIndexingInterface.state_values — Function

state_values(valp)
state_values(valp, i)

Return an indexable collection containing the values of all states in the value provider p. If is_timeseries(valp) is Timeseries, return a vector of arrays, each of which contain the state values at the corresponding timestep. In this case, the two-argument version of the function can also be implemented to efficiently return the state values at timestep i. By default, the two-argument method calls state_values(valp)[i]

If this function is called with an AbstractArray, it will return the same array.

See: is_timeseries

SymbolicIndexingInterface.set_state! — Function

set_state!(valp, val, idx)

Set the state at index idx to val for value provider valp. This defaults to modifying state_values(valp). If any additional bookkeeping needs to be performed or the default implementation does not work for a particular type, this method needs to be defined to enable the proper functioning of setu.

See: state_values

SymbolicIndexingInterface.current_time — Function

current_time(valp)
current_time(valp, i)

Return the current time in the value provider valp. If is_timeseries(valp) is Timeseries, return the vector of timesteps at which the state value is saved. In this case, the two-argument version of the function can also be implemented to efficiently return the time at timestep i. By default, the two- argument method calls current_time(p)[i]

See: is_timeseries

SymbolicIndexingInterface.getu — Function

getu(indp, sym)

Return a function that takes a value provider and returns the value of the symbolic variable sym. If sym is not an observed quantity, the returned function can also directly be called with an array of values representing the state vector. sym can be an index into the state vector, a symbolic variable, a symbolic expression involving symbolic variables in the index provider indp, a parameter symbol, the independent variable symbol, or an array/tuple of the aforementioned. If the returned function is called with a timeseries object, it can also be given a second argument representing the index at which to return the value of sym.

At minimum, this requires that the value provider implement state_values. To support symbolic expressions, the value provider must implement observed, parameter_values and current_time.

This function typically does not need to be implemented, and has a default implementation relying on the above functions.

getu(bi::BatchedInterface)

Given a BatchedInterface composed from n index providers (and corresponding symbols), return a function which takes n corresponding value providers and returns an array of the values of the symbols in the union. The returned function can also be passed an AbstractArray of the appropriate eltype and size as its first argument, in which case the operation will populate the array in-place with the values of the symbols in the union.

Note that all of the value providers passed to the function returned by getu must satisfy is_timeseries(prob) === NotTimeseries().

The value of the ith symbol in the union (obtained through variable_symbols(bi)[i]) is obtained from the problem corresponding to the associated index provider (i.e. the value provider at index associated_systems(bi)[i]).

See also: variable_symbols, associated_systems, is_timeseries, NotTimeseries.

SymbolicIndexingInterface.setu — Function

setu(sys, sym)

Return a function that takes a value provider and a value, and sets the the state sym to that value. Note that sym can be an index, a symbolic variable, or an array/tuple of the aforementioned.

Requires that the value provider implement state_values and the returned collection be a mutable reference to the state vector in the value provider. Alternatively, if this is not possible or additional actions need to be performed when updating state, set_state! can be defined. This function does not work on types for which is_timeseries is Timeseries.

setu(bi::BatchedInterface)

Given a BatchedInterface composed from n index providers (and corresponding symbols), return a function which takes n corresponding problems and an array of the values, and updates each of the problems with the values of the corresponding symbols.

Note that all of the value providers passed to the function returned by setu must satisfy is_timeseries(prob) === NotTimeseries().

Note that if any subset of the n index providers share common symbols (among those passed to BatchedInterface) then all of the corresponding value providers in the subset will be updated with the values of the common symbols.

See also: is_timeseries, NotTimeseries.

Batched Queries and Updates

SymbolicIndexingInterface.BatchedInterface — Type

struct BatchedInterface{S <: AbstractVector, I}
function BatchedInterface(indp_syms::Tuple...)

A struct which stores information for batched calls to getu or setu. Given Tuples, where the first element of each tuple is an index provider and the second an array of symbolic variables (either states or parameters) in the index provider, BatchedInterface will compute the union of all symbols and associate each symbol with the first index provider with which it occurs.

For example, given two index providers s1 = SymbolCache([:x, :y, :z]) and s2 = SymbolCache([:y, :z, :w]), BatchedInterface((s1, [:x, :y]), (s2, [:y, :z])) will associate :x and :y with s1 and :z with s2. The information that s1 had associated symbols :x and :y and s2 had associated symbols :y and :z will also be retained internally.

BatchedInterface implements variable_symbols, is_variable, variable_index to query the order of symbols in the union.

See getu and setu for further details.

See also: associated_systems.

SymbolicIndexingInterface.associated_systems — Function

associated_systems(bi::BatchedInterface)

Return an array of integers of the same length as variable_symbols(bi) where each value is the index of the index provider associated with the corresponding symbol in variable_symbols(bi).

Container objects

SymbolicIndexingInterface.remake_buffer — Function

remake_buffer(indp, oldbuffer, vals::Dict)

Return a copy of the buffer oldbuffer with values from vals. The keys of vals are symbolic variables whose index in the buffer is determined using indp. The types of values in vals may not match the types of values stored at the corresponding indexes in the buffer, in which case the type of the buffer should be promoted accordingly. In general, this method should attempt to preserve the types of values stored in vals as much as possible. Types can be promoted for type-stability, to maintain performance. The returned buffer should be of the same type (ignoring type-parameters) as oldbuffer.

This method is already implemented for remake_buffer(indp, oldbuffer::AbstractArray, vals::Dict) and supports static arrays as well. It is also implemented for oldbuffer::Tuple.

Symbolic Trait

SymbolicIndexingInterface.ScalarSymbolic — Type

struct ScalarSymbolic <: SymbolicTypeTrait end

Trait indicating a type is a scalar symbolic variable.

See also: ArraySymbolic, NotSymbolic, symbolic_type

SymbolicIndexingInterface.ArraySymbolic — Type

struct ArraySymbolic <: SymbolicTypeTrait end

Trait indicating type is a symbolic array. Calling collect on a symbolic array must return an AbstractArray containing ScalarSymbolic variables for each element in the array, in the same shape as the represented array. For example, if a is a symbolic array representing a 2x2 matrix, collect(a) must return a 2x2 array of scalar symbolic variables.

See also: ScalarSymbolic, NotSymbolic, symbolic_type

SymbolicIndexingInterface.NotSymbolic — Type

struct NotSymbolic <: SymbolicTypeTrait end

Trait indicating a type is not symbolic.

See also: ScalarSymbolic, ArraySymbolic, symbolic_type

SymbolicIndexingInterface.symbolic_type — Function

symbolic_type(x) = symbolic_type(typeof(x))
symbolic_type(::Type)

Get the symbolic type trait of a type. Default to NotSymbolic for all types except Symbol and Expr, both of which are ScalarSymbolic.

See also: ScalarSymbolic, ArraySymbolic, NotSymbolic

SymbolicIndexingInterface.hasname — Function

hasname(x)

Check whether the given symbolic variable (for which symbolic_type(x) != NotSymbolic()) has a valid name as per getname. Defaults to true for x::Symbol.

SymbolicIndexingInterface.getname — Function

getname(x)::Symbol

Get the name of a symbolic variable as a Symbol. Acts as the identity function for x::Symbol.

SymbolicIndexingInterface.symbolic_evaluate — Function

symbolic_evaluate(expr, syms::Dict; kwargs...)

Return the value of symbolic expression expr where the values of variables involved are obtained from the dictionary syms. The keys of syms are symbolic variables (not expressions of variables). The values of syms can be values or symbolic expressions.

The returned value should either be a value or an expression involving symbolic variables not present as keys in syms.

The function can take additional keyword arguments to control implementation-specific behavior.

This is already implemented for symbolic_evaluate(expr::Union{Symbol, Expr}, syms::Dict).

Types

SymbolicIndexingInterface.SymbolCache — Type

struct SymbolCache{V,P,I}
function SymbolCache(vars, [params, [indepvars]])

A struct implementing the index provider interface for the trivial case of having a vector of variables, parameters, and independent variables. It is considered time dependent if it contains at least one independent variable. It returns true for is_observed(::SymbolCache, sym) if sym isa Expr. Functions can be generated using observed for Exprs involving variables in the SymbolCache if it has at most one independent variable.

The independent variable may be specified as a single symbolic variable instead of an array containing a single variable if the system has only one independent variable.

SymbolicIndexingInterface.ProblemState — Type

struct ProblemState
function ProblemState(; u = nothing, p = nothing, t = nothing)

A value provider struct which can be used as an argument to the function returned by getu or setu. It stores the state vector, parameter object and current time, and forwards calls to state_values, parameter_values, current_time, set_state!, set_parameter! to the contained objects.