Interface Functions

SymbolicIndexingInterface.isvariable — Function

isvariable(sys, sym)

Check whether the given sym is a variable in sys.

SymbolicIndexingInterface.variableindex — Function

variableindex(sys, sym)

Return the index of the given variable sym in sys, or nothing otherwise.

SymbolicIndexingInterface.isparameter — Function

isparameter(sys, sym)

Check whether the given sym is a parameter in sys.

SymbolicIndexingInterface.parameterindex — Function

parameterindex(sys, sym)

Return the index of the given parameter sym in sys, or nothing otherwise.

SymbolicIndexingInterface.isindependent_variable — Function

isindependent_variable(sys, sym)

Check whether the given sym is an independent variable in sys.

SymbolicIndexingInterface.isobserved — Function

isobserved(sys, sym)

Check whether the given sym is an observed value in sys.

SymbolicIndexingInterface.observed — Function

observed(sys, sym)

Return the observed function of the given sym in sys. The returned function should accept a timeseries if sys has an independent variable, and return the observed values for the given sym.

SymbolicIndexingInterface.istimedependent — Function

istimedependent(sys)

Check if sys has time as (one of) its independent variables.

SymbolicIndexingInterface.constant_structure — Function

constant_structure(sys)

Check if sys has a constant structure. Constant structure systems do not change the number of variables or parameters over time.