Interface Functions

Default methods cast all symbols to Symbol before comparing.

SymbolicIndexingInterface.independent_variables — Function

Get an iterable over the independent variables for the given system. Default to an empty vector.

SymbolicIndexingInterface.is_indep_sym — Function

Check if the given sym is an independent variable in the given system. Default to checking if the given sym exists in the iterable returned by independent_variables.

SymbolicIndexingInterface.states — Function

Get an iterable over the states for the given system. Default to an empty vector.

SymbolicIndexingInterface.state_sym_to_index — Function

Find the index of the given sym in the given system. Default to the index of the first symbol in the iterable returned by states which matches the given sym. Return nothing if the given sym does not match.

SymbolicIndexingInterface.is_state_sym — Function

Check if the given sym is a state variable in the given system. Default to checking if the value returned by state_sym_to_index is not nothing.

SymbolicIndexingInterface.parameters — Function

Get an iterable over the parameters variables for the given system. Default to an empty vector.

SymbolicIndexingInterface.param_sym_to_index — Function

Find the index of the given sym in the given system. Default to the index of the first symbol in the iterable retruned by parameters which matches the given sym. Return nothing if the given sym does not match.

SymbolicIndexingInterface.is_param_sym — Function

Check if the given sym is a parameter variable in the given system. Default to checking if the value returned by param_sym_to_index is not nothing.

Concrete Types

SymbolicIndexingInterface.SymbolCache — Type

SymbolCache(syms, indepsym, paramsyms)

A container that simply stores a vector of all syms, indepsym and paramsyms.